refactor: review comments handled
