+++ src/modules/testimonials/dto/create-testimonial-response.dto.ts
@@ -1,0 +1,7 @@
+import { TestimonialData } from '../interfaces/testimonials.interface';
+
+export class CreateTestimonialResponseDto {
+  'status': string;
+  'message': string;
+  'data': TestimonialData;
+}
