+++ src/modules/testimonials/interfaces/testimonials.interface.ts
@@ -1,0 +1,6 @@
+export interface TestimonialData {
+  user_id: string;
+  name: string;
+  content: string;
+  created_at: Date;
+}
