+++ src/modules/testimonials/testimonials.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from '../user/entities/user.entity';
+import UserService from '../user/user.service';
 import { Testimonial } from './entities/testimonials.entity';
 import { TestimonialsController } from './testimonials.controller';
 import { TestimonialsService } from './testimonials.service';
@@ -9,6 +10,6 @@ import { TestimonialsService } from './testimonials.service';
 @Module({
   imports: [TypeOrmModule.forFeature([Testimonial, User])],
   controllers: [TestimonialsController],
-  providers: [TestimonialsService, Repository],
+  providers: [TestimonialsService, Repository, UserService],
 })
 export class TestimonialsModule {}
