+++ src/modules/testimonials/testimonials.service.ts
@@ -1,7 +1,6 @@
-import { HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { User } from '../user/entities/user.entity';
 import { CreateTestimonialDto } from './dto/create-testimonial.dto';
 import { Testimonial } from './entities/testimonials.entity';
 
@@ -9,23 +8,10 @@ import { Testimonial } from './entities/testimonials.entity';
 export class TestimonialsService {
   constructor(
     @InjectRepository(Testimonial)
-    private readonly testimonialRepository: Repository<Testimonial>,
-
-    @InjectRepository(User)
-    private readonly userRepository: Repository<User>
+    private readonly testimonialRepository: Repository<Testimonial>
   ) {}
-  async create(createTestimonialDto: CreateTestimonialDto, userId: string) {
+  async createTestimonial(createTestimonialDto: CreateTestimonialDto, user) {
     const { content, name } = createTestimonialDto;
-
-    const user = await this.userRepository.findOneBy({ id: userId });
-
-    if (!user) {
-      throw new NotFoundException({
-        status: 'Not Found',
-        message: 'User not found',
-        status_code: HttpStatus.NOT_FOUND,
-      });
-    }
 
     await this.testimonialRepository.save({
       user,
